Move client only command into new package
