+++ src/main/java/com/sinthoras/visualprospecting/commands/ResetClientCacheCommand.java
@@ -1,4 +1,4 @@
-package com.sinthoras.visualprospecting.database;
+package com.sinthoras.visualprospecting.commands;
 
 import net.minecraft.client.resources.I18n;
 import net.minecraft.command.CommandBase;
@@ -6,23 +6,25 @@ import net.minecraft.command.ICommandSender;
 import net.minecraft.util.ChatComponentTranslation;
 import net.minecraft.util.IChatComponent;
 
+import com.sinthoras.visualprospecting.database.ClientCache;
+
 public class ResetClientCacheCommand extends CommandBase {
 
     @Override
     public String getCommandName() {
-        return "visualprospectingresetprogress";
+        return "vp_client_cache_reset";
     }
 
     @Override
     public String getCommandUsage(ICommandSender sender) {
-        return I18n.format("visualprospecting.resetprogress.command");
+        return I18n.format("visualprospecting.command.client.resetcache.usage");
     }
 
     @Override
     public void processCommand(ICommandSender sender, String[] parameters) {
         ClientCache.instance.resetPlayerProgression();
         final IChatComponent confirmation = new ChatComponentTranslation(
-                "visualprospecting.resetprogress.confirmation");
+                "visualprospecting.command.client.resetcache.confirmation");
         confirmation.getChatStyle().setItalic(true);
         sender.addChatMessage(confirmation);
     }
