+++ src/main/java/com/sinthoras/visualprospecting/hooks/HooksClient.java
@@ -6,8 +6,8 @@ import net.minecraftforge.common.MinecraftForge;
 import com.gtnewhorizons.navigator.api.NavigatorApi;
 import com.gtnewhorizons.navigator.api.util.Util;
 import com.sinthoras.visualprospecting.Utils;
+import com.sinthoras.visualprospecting.commands.ResetClientCacheCommand;
 import com.sinthoras.visualprospecting.database.ClientCache;
-import com.sinthoras.visualprospecting.database.ResetClientCacheCommand;
 import com.sinthoras.visualprospecting.integration.model.layers.OreVeinLayerManager;
 import com.sinthoras.visualprospecting.integration.model.layers.UndergroundFluidLayerManager;
 import com.sinthoras.visualprospecting.integration.voxelmap.VoxelMapEventHandler;
