+++ src/main/resources/assets/visualprospecting/lang/de_DE.lang
@@ -6,8 +6,6 @@ visualprospecting.undergroundfluid.prospected.onlyupdated=Du hast %s Ölfelder v
 visualprospecting.undergroundfluid.prospected.newandupdated=Du hast %s Ölfinder neu gefunden und %s vermessen!
 visualprospecting.button.orevein=GT-Erzadern anzeigen
 visualprospecting.button.undergroundfluid=GT-Gas/Flüssigkeitsadern anzeigen
-visualprospecting.resetprogress.command=Löscht deinen gesamten Fortschritt beim Entdecken von Erz- und Flüssigkeitsadern
-visualprospecting.resetprogress.confirmation=Fortschritt gelöscht!
 visualprospecting.redoservercache.command=GT-Ader-Cache auf dem Server wiederherstellen
 visualprospecting.redoservercache.start=Cache-wiederherstellen auf dem Server beginnen
 visualprospecting.redoservercache.failure=Konnte Weltdaten nicht laden, um Cache wiederherstellen!
@@ -40,3 +38,5 @@ item.visualprospecting.prospectorslog.reading.begin=Das ist mal ein Wälzer. Das
 item.visualprospecting.prospectorslog.reading.end=Goldsucher-Tagebuch von %s durchgelesen
 visualprospecting.node.deletehint=Mit %s als erschöpft markieren
 visualprospecting.tracked=Nachverfolgt: %s
+visualprospecting.command.client.resetcache.usage=Löscht deinen gesamten Fortschritt beim Entdecken von Erz- und Flüssigkeitsadern
+visualprospecting.command.client.resetcache.confirmation=Fortschritt gelöscht!
