+++ src/main/resources/assets/visualprospecting/lang/en_US.lang
@@ -6,8 +6,6 @@ visualprospecting.undergroundfluid.prospected.onlyupdated=You updated %s oil fie
 visualprospecting.undergroundfluid.prospected.newandupdated=You found %s new and updated %s oil fields!
 visualprospecting.button.orevein=Show GT Ore Veins
 visualprospecting.button.undergroundfluid=Show GT Underground Fluids
-visualprospecting.resetprogress.command=Deletes all your progress on ore veins and underground fluids
-visualprospecting.resetprogress.confirmation=Progress deleted!
 visualprospecting.redoservercache.command=Redo the GT vein cache on the server
 visualprospecting.redoservercache.start=Start redoing the GT vein cache on the server
 visualprospecting.redoservercache.failure=Could not load world save files to build vein cache!
@@ -41,6 +39,8 @@ item.visualprospecting.prospectorslog.reading.end=I finished the Prospector's Lo
 visualprospecting.node.deletehint=Mark depleted with %s
 visualprospecting.tracked=Tracked: %s
 visualprospecting.command.usage_error=Invalid arguments for /%s.
+visualprospecting.command.client.resetcache.usage=/vp_client_cache_reset: Deletes local cache of all your ore veins and underground fluids
+visualprospecting.command.client.resetcache.confirmation=Local cache deleted. It may still reconstruct partially from team cache on next logon if they are enabled.
 visualprospecting.teamcommand.console_needs_player=Must specify a player when running command from console.
 visualprospecting.teamcommand.player_offline=Player not found or not online: %s
 visualprospecting.teamcommand.no_team=%s is not in any team.
