+++ src/main/resources/assets/visualprospecting/lang/zh_CN.lang
@@ -6,8 +6,6 @@ visualprospecting.undergroundfluid.prospected.onlyupdated=你更新了 %s 流体
 visualprospecting.undergroundfluid.prospected.newandupdated=你找到了 %s 新流体井并更新了 %s 流体井!
 visualprospecting.button.orevein=显示GT矿脉
 visualprospecting.button.undergroundfluid=显示GT地下流体
-visualprospecting.resetprogress.command=删除你的所有矿脉与地下流体记录
-visualprospecting.resetprogress.confirmation=记录已删除!
 visualprospecting.redoservercache.command=在服务器上重新缓存GT矿脉
 visualprospecting.redoservercache.start=开始重新缓存GT矿脉
 visualprospecting.redoservercache.failure=无法加载世界存档以构建矿脉缓存
@@ -33,3 +31,5 @@ item.visualprospecting.prospectorslog.reading.begin=有相当多页.可能得多
 item.visualprospecting.prospectorslog.reading.end=我读完了 %s 的探矿记录
 visualprospecting.node.deletehint=删除 %s
 visualprospecting.tracked=追踪: %s
+visualprospecting.command.client.resetcache.usage=删除你的所有矿脉与地下流体记录
+visualprospecting.command.client.resetcache.confirmation=记录已删除!
